requirements.txt
defusedxml==0.6.0
Django==2.2.2
coverage
django-tastypie==0.14.2
djangorestframework==3.9.4
drf-writable-nested==0.5.1
factory-boy==2.12.0
Faker==1.0.7
filelock==3.0.12
importlib-metadata==0.18
lxml==4.3.4
pluggy==0.12.0
py==1.8.0
python-dateutil==2.8.0
python-mimeparse==1.6.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
toml==0.10.0
tox==3.12.1
virtualenv==16.6.1
zipp==0.5.1
